feat(resource): async guard release, pool maintenance reaper, Manager split
Three related lifecycle/structure changes to nebula-resource:

* Explicit `ResourceGuard::release(self) -> Result<(), Error>` — the awaited
  release checkpoint canon §11.4 calls for. Release callbacks now return the
  teardown future instead of submitting it; `Drop` submits-and-discards
  (best-effort, byte-equivalent behavior), `release()` awaits it inline and
  surfaces the recycle/destroy/reset result. Drain accounting always settles
  (even on Err), #384 permit-held-until-reset and bounded S4 poison are
  preserved, and the consumed guard is neutralized so Drop cannot double-settle.

* Background pool maintenance reaper — `PoolRuntime::run_maintenance` (idle /
  max-lifetime / stale-fingerprint / revoked eviction) had no production
  driver, so `idle_timeout` was dead config. `Manager::register` now spawns a
  reaper per Pooled registration when a TTL is set, holding a Weak to the
  ManagedResource (no lifetime extension) and the manager cancel token (stops
  on shutdown; self-terminates on remove). Adds
  `ResourceEvent::MaintenanceEvicted` for observability;
  `run_maintenance`/`should_evict` relaxed to `R: Resource`.

* Manager god-file split — the single ~2100-line `impl Manager` block in
  manager/mod.rs (2991 -> 1149 LoC) is carved into concern-named sibling
  `impl Manager` blocks: acquire.rs, registration.rs, rotation.rs. Pure move;
  public API unchanged (three private helpers elevated to pub(crate) for the
  in-mod test module).

Verified: clippy -D warnings clean, 343 nextest pass (incl. revoke-TOCTOU /
S4 / #384 / shutdown-race + new release/reaper tests), rustdoc -D warnings
clean. Adversarial Rust reviews on the split and both features found no
correctness defects.
