+++ crates/resource/src/manager/rotation.rs
@@ -1,0 +1,600 @@
+//! Credential slot rotation / revoke: the refresh + two-phase
+//! (synchronous-taint then cancellation-safe drain+hook) revoke surface,
+//! their shared post-resolution dispatch, and the type-erased `(key, scope)`
+//! row resolution helpers the rotation entry points use.
+
+use std::{sync::Arc, time::Instant};
+
+use nebula_core::{ResourceKey, ScopeLevel};
+
+use super::{Manager, RevokeTail, TaintedSlot};
+use crate::{error::Error, events::ResourceEvent};
+
+impl Manager {
+    /// Notifies a registered resource that one of its `#[credential]`
+    /// slots was rotated, after the engine has installed the fresh guard.
+    ///
+    /// Resolves `(key, scope)` to the live [`ManagedResource`](crate::ManagedResource) via the same
+    /// registry lookup the `acquire_*` family uses, then borrows the live
+    /// `Runtime` per topology and invokes
+    /// [`Resource::on_credential_refresh`](crate::resource::Resource::on_credential_refresh)
+    /// for `slot`. The slot cell itself
+    /// lives on the author's resource struct and is populated/rotated by
+    /// the engine through `&self` (`SlotCell::store`) — this method does
+    /// **not** own a slot map; it only drives the per-resource hook.
+    ///
+    /// Emits [`ResourceEvent::SlotRefreshed`] on success or
+    /// [`ResourceEvent::SlotRefreshFailed`] (with an already-stringified,
+    /// credential-free error) on failure, and records the corresponding
+    /// slot-refresh metric.
+    ///
+    /// # Errors
+    ///
+    /// - [`ErrorKind::NotFound`](crate::error::ErrorKind::NotFound) if no resource is registered for
+    ///   `key` at `scope`.
+    /// - [`ErrorKind::Cancelled`](crate::error::ErrorKind::Cancelled) if the manager is shutting
+    ///   down.
+    /// - Whatever the resource's `on_credential_refresh` hook maps into [`Error`].
+    #[tracing::instrument(
+        level = "debug",
+        name = "nebula.resource.slot_refresh",
+        skip(self),
+        fields(key = %key, slot = %slot, topology, duration_ms)
+    )]
+    pub async fn refresh_slot(
+        &self,
+        key: &ResourceKey,
+        scope: ScopeLevel,
+        slot: &str,
+    ) -> Result<(), Error> {
+        let managed = self.lookup_any_for_slot(key, &scope)?;
+        self.refresh_resolved(key, slot, managed).await
+    }
+
+    /// [`refresh_slot`](Self::refresh_slot) pinned to the **collision-free
+    /// structural** resolved per-slot credential identity.
+    ///
+    /// Resolves the registry row whose `slot_identity` matches (via the same
+    /// unambiguous-by-construction path [`get_for`](crate::registry::Registry::get_for)
+    /// backs), so a multi-tenant `(key, scope)` routes the rotation to the
+    /// *specific* resolved row instead of failing closed with
+    /// [`ErrorKind::Ambiguous`](crate::error::ErrorKind::Ambiguous). This is
+    /// the entry point the engine per-slot rotation fan-out drives once it
+    /// has resolved a node's slot bindings; identity-agnostic
+    /// [`refresh_slot`](Self::refresh_slot) stays fail-closed for the
+    /// no-identity caller. The engine rotation fan-out records the
+    /// structural [`SlotIdentity`](crate::dedup::SlotIdentity) at bind time,
+    /// so routing is by exact string equality (no digest aliasing).
+    ///
+    /// # Errors
+    ///
+    /// - [`ErrorKind::NotFound`](crate::error::ErrorKind::NotFound) if no row of `key` at `scope`
+    ///   matches `slot_identity`.
+    /// - [`ErrorKind::Cancelled`](crate::error::ErrorKind::Cancelled) if the manager is shutting
+    ///   down.
+    /// - Whatever the resource's `on_credential_refresh` hook maps into [`Error`].
+    #[tracing::instrument(
+        level = "debug",
+        name = "nebula.resource.slot_refresh",
+        skip(self, slot_identity),
+        fields(key = %key, slot = %slot, topology, duration_ms)
+    )]
+    pub async fn refresh_slot_for_identity(
+        &self,
+        key: &ResourceKey,
+        scope: ScopeLevel,
+        slot: &str,
+        slot_identity: &crate::dedup::SlotIdentity,
+    ) -> Result<(), Error> {
+        let managed = self.lookup_any_for_slot_identity_structural(key, &scope, slot_identity)?;
+        self.refresh_resolved(key, slot, managed).await
+    }
+
+    /// Post-resolution refresh dispatch shared by
+    /// [`refresh_slot`](Self::refresh_slot) (identity-agnostic) and
+    /// [`refresh_slot_for_identity`](Self::refresh_slot_for_identity)
+    /// (slot-identity-pinned).
+    ///
+    /// The two public entry points differ only in how they resolve the row;
+    /// the hook dispatch, metric (exactly one outcome per dispatch), and
+    /// event emission are identical and live here.
+    async fn refresh_resolved(
+        &self,
+        key: &ResourceKey,
+        slot: &str,
+        managed: Arc<dyn crate::registry::AnyManagedResource>,
+    ) -> Result<(), Error> {
+        let started = Instant::now();
+        tracing::Span::current().record("topology", managed.topology_tag_erased().as_str());
+
+        let result = managed.dispatch_on_refresh_erased(slot).await;
+        tracing::Span::current().record("duration_ms", started.elapsed().as_millis() as u64);
+
+        // Exactly one outcome per dispatch; the attempts total is the sum
+        // across `outcome` labels (success + failed + timed_out).
+        match &result {
+            Ok(()) => {
+                if let Some(m) = &self.metrics {
+                    m.record_slot_refresh_outcome(crate::metrics::SlotDispatchOutcome::Success);
+                }
+                self.emit(ResourceEvent::SlotRefreshed {
+                    key: key.clone(),
+                    slot: slot.to_owned(),
+                });
+                tracing::debug!("slot refresh hook completed");
+            },
+            Err(e) => {
+                if let Some(m) = &self.metrics {
+                    m.record_slot_refresh_outcome(crate::metrics::SlotDispatchOutcome::Failed);
+                }
+                self.emit(ResourceEvent::SlotRefreshFailed {
+                    key: key.clone(),
+                    slot: slot.to_owned(),
+                    error: e.to_string(),
+                });
+                tracing::warn!(error = %e, "slot refresh hook failed");
+            },
+        }
+        result
+    }
+
+    /// **Phase 1 of the revoke port — synchronous, runs to completion before
+    /// any `.await`.** Resolves the registry row pinned to the
+    /// **collision-free structural** resolved per-slot credential identity
+    /// and *taints it immediately* so the `acquire_*` funnel rejects new
+    /// leases on the revoked credential, then returns a [`TaintedSlot`]
+    /// handle the caller passes to
+    /// [`drain_and_revoke`](Self::drain_and_revoke) for the cancellation-safe
+    /// drain + hook tail.
+    ///
+    /// Why this is split off as a non-`async` function: the engine fan-out
+    /// wraps the awaited tail in `tokio::time::timeout`. A Rust `async fn`
+    /// body is *lazy* — if a `timeout` future is dropped before the runtime
+    /// first polls it, the body never runs. Were the taint the first
+    /// statement of an `async` body, a timeout that fired before the first
+    /// poll would drop the future and **skip the taint entirely**, leaving
+    /// new acquires accepted on a credential whose revoke "timed out". This
+    /// function is plain `fn`: the taint is applied eagerly at the call site,
+    /// fully completed before this returns, and therefore *outside* and
+    /// *before* any per-resource timeout (per-resource revoke deferral).
+    ///
+    /// Identity routing: resolves the *exact* resolved registry row by
+    /// structural string equality (no digest aliasing) via the
+    /// unambiguous-by-construction
+    /// [`get_for`](crate::registry::Registry::get_for) path, so a
+    /// multi-tenant `(key, scope)` taints the *specific* resolved row
+    /// instead of failing closed with
+    /// [`ErrorKind::Ambiguous`](crate::error::ErrorKind::Ambiguous). This is
+    /// the entry point the engine per-slot rotation fan-out drives on a
+    /// lease revoke; identity-agnostic [`taint_slot`](Self::taint_slot) stays
+    /// fail-closed for the no-identity caller. Synchronous-before-`.await`
+    /// taint guarantee; see the [`manager`](crate::manager) module docs for
+    /// the canonical invariant.
+    ///
+    /// # Errors
+    ///
+    /// - [`ErrorKind::NotFound`](crate::error::ErrorKind::NotFound) if no row of `key` at `scope`
+    ///   matches `slot_identity`.
+    /// - [`ErrorKind::Cancelled`](crate::error::ErrorKind::Cancelled) if the manager is shutting
+    ///   down.
+    ///
+    /// Carries only `key` / `slot` / `topology` (no credential material)
+    /// onto the span.
+    #[tracing::instrument(
+        level = "debug",
+        name = "nebula.resource.slot_taint",
+        skip(self, slot_identity),
+        fields(key = %key, slot = %slot, topology, op = "revoke")
+    )]
+    pub fn taint_slot_for_identity(
+        &self,
+        key: &ResourceKey,
+        scope: ScopeLevel,
+        slot: &str,
+        slot_identity: &crate::dedup::SlotIdentity,
+    ) -> Result<TaintedSlot, Error> {
+        let managed = self.lookup_any_for_slot_identity_structural(key, &scope, slot_identity)?;
+        Ok(Self::taint_now(key, slot, managed))
+    }
+
+    /// [`taint_slot_for_identity`](Self::taint_slot_for_identity) for the
+    /// slot-identity-agnostic caller (the convenience
+    /// [`revoke_slot`](Self::revoke_slot) path and non-fan-out
+    /// callers/tests).
+    ///
+    /// Same eager, pre-`await` taint guarantee as
+    /// [`taint_slot_for_identity`](Self::taint_slot_for_identity); only row
+    /// resolution differs (identity-agnostic, so a multi-tenant
+    /// `(key, scope)` fails closed with
+    /// [`ErrorKind::Ambiguous`](crate::error::ErrorKind::Ambiguous) rather
+    /// than tainting an arbitrary tenant's row).
+    ///
+    /// # Errors
+    ///
+    /// - [`ErrorKind::NotFound`](crate::error::ErrorKind::NotFound) if no resource is registered for
+    ///   `key` at `scope`.
+    /// - [`ErrorKind::Ambiguous`](crate::error::ErrorKind::Ambiguous) if more than one
+    ///   resolved-credential row exists for `(key, scope)`.
+    /// - [`ErrorKind::Cancelled`](crate::error::ErrorKind::Cancelled) if the manager is shutting
+    ///   down.
+    #[tracing::instrument(
+        level = "debug",
+        name = "nebula.resource.slot_taint",
+        skip(self),
+        fields(key = %key, slot = %slot, topology, op = "revoke")
+    )]
+    pub fn taint_slot(
+        &self,
+        key: &ResourceKey,
+        scope: ScopeLevel,
+        slot: &str,
+    ) -> Result<TaintedSlot, Error> {
+        let managed = self.lookup_any_for_slot(key, &scope)?;
+        Ok(Self::taint_now(key, slot, managed))
+    }
+
+    /// Applies the taint synchronously and packages the [`TaintedSlot`]
+    /// handle. Shared tail of [`taint_slot`](Self::taint_slot) /
+    /// [`taint_slot_for_identity`](Self::taint_slot_for_identity); the
+    /// safety-critical
+    /// invariant — *taint is fully applied before this returns* — is written
+    /// once here. This is **phase 1** of the two-phase revoke; see the
+    /// [`manager`](crate::manager) module docs for the canonical invariant
+    /// (why both stores are synchronous-before-`.await`, the TOCTOU close,
+    /// and the revoke-epoch fence).
+    fn taint_now(
+        key: &ResourceKey,
+        slot: &str,
+        managed: Arc<dyn crate::registry::AnyManagedResource>,
+    ) -> TaintedSlot {
+        tracing::Span::current().record("topology", managed.topology_tag_erased().as_str());
+        // Phase-1 taint, synchronously before any caller `.await`: this
+        // function is not `async`, so the store has already happened by the
+        // time control returns and a subsequently-dropped drain-tail timeout
+        // future cannot un-apply it.
+        managed.taint_erased();
+        // Phase-1 revoke-epoch bump, in the *same* synchronous pre-`.await`
+        // step as the taint, so the pooled return-to-idle paths fence any
+        // instance authenticated with the now-revoked credential before the
+        // hook walks the idle queue.
+        managed.bump_revoke_epoch_erased();
+        TaintedSlot {
+            key: key.clone(),
+            slot: slot.to_owned(),
+            managed,
+            tainted_at: Instant::now(),
+        }
+    }
+
+    /// Default per-resource revoke budget for the back-compat
+    /// back-to-back convenience callers ([`revoke_slot`](Self::revoke_slot)
+    /// / [`revoke_slot_for_identity`](Self::revoke_slot_for_identity)).
+    ///
+    /// 30 s — the same budget the manager-wide `graceful_shutdown` drain
+    /// uses and the value [`drain_and_revoke`](Self::drain_and_revoke)
+    /// previously hard-coded for the drain wait. The engine rotation
+    /// fan-out does **not** use this: it passes its own per-resource
+    /// rotation budget so the timeout has one owner end-to-end (resource runtime status
+    /// §Deferred / #690 review).
+    pub const DEFAULT_REVOKE_DRAIN_TIMEOUT: std::time::Duration =
+        std::time::Duration::from_secs(30);
+
+    /// **Phase 2 of the revoke port — the cancellation-safe awaited tail.**
+    /// Consumes a [`TaintedSlot`] from [`taint_slot`](Self::taint_slot) /
+    /// [`taint_slot_for_identity`](Self::taint_slot_for_identity) (whose
+    /// taint already ran
+    /// synchronously) and performs the remaining steps:
+    ///
+    /// 1. **Drain** only *this resource's* in-flight handles via its own per-resource counter
+    ///    (per-resource revoke deferral) — never the manager-wide `drain_tracker`, so a revoke is isolated
+    ///    from in-flight traffic to unrelated resources.
+    /// 2. **Dispatch** [`Resource::on_credential_revoke`](crate::resource::Resource::on_credential_revoke) against the live runtime per topology.
+    /// 3. Emit [`ResourceEvent::SlotRevoked`] / `SlotRevokeFailed`.
+    ///
+    /// **Single budget owner (per-resource revoke deferral / #690 review).** The
+    /// `drain_timeout` argument is the caller's per-resource budget and is
+    /// the *only* timeout governing this tail. It bounds **two** waits
+    /// independently:
+    ///
+    /// - the per-resource **drain** — *best-effort*: a drain timeout is
+    ///   non-fatal, it records the `TimedOut` outcome metric and the tail
+    ///   **still proceeds to the revoke hook** (the taint already stops
+    ///   *new* leases; the hook makes the resource stop emitting on the
+    ///   old credential);
+    /// - the **revoke hook** itself — a *wedged* `on_credential_revoke`
+    ///   is the only thing the budget actually cuts short
+    ///   ([`RevokeTail::HookTimedOut`]).
+    ///
+    /// The caller **must not** wrap this call in its own
+    /// `tokio::time::timeout`. The previous design did, and a slow drain
+    /// could make that outer timeout elapse and **drop the whole future
+    /// before the hook ran** — silently skipping the documented
+    /// "hook still runs after a timed-out drain" guarantee. Bounding both
+    /// waits *inside* this method (one owner, no outer wrapper) means a
+    /// timed-out drain can never skip the hook, and only a hung hook is
+    /// bounded — never the taint.
+    ///
+    /// **Cancellation-safety.** The taint is *not* in this future — it
+    /// ran in the synchronous
+    /// [`taint_slot_for_identity`](Self::taint_slot_for_identity)
+    /// phase. So if this future *is* dropped anyway (an outer abort, task
+    /// cancel), the row stays tainted and consistent: new acquires are
+    /// still rejected, the credential is never silently un-revoked.
+    #[tracing::instrument(
+        level = "debug",
+        name = "nebula.resource.slot_drain_revoke",
+        skip(self, tainted),
+        fields(
+            key = %tainted.key,
+            slot = %tainted.slot,
+            topology = tainted.managed.topology_tag_erased().as_str(),
+            duration_ms,
+            op = "revoke",
+        )
+    )]
+    pub async fn drain_and_revoke(
+        &self,
+        tainted: TaintedSlot,
+        drain_timeout: std::time::Duration,
+    ) -> RevokeTail {
+        let TaintedSlot {
+            key,
+            slot,
+            managed,
+            tainted_at,
+        } = tainted;
+
+        // 1. Drain **only this resource's** in-flight handles (resource runtime status
+        //    §Deferred): a revoke on resource A must not block on in-flight
+        //    traffic to an unrelated resource B, so this awaits the row's
+        //    own per-resource counter — not the manager-wide `drain_tracker`
+        //    (which stays the `graceful_shutdown` primitive). Bounded by the
+        //    caller's per-resource budget so a stuck handle on *this*
+        //    resource cannot wedge revoke; the taint (already applied
+        //    synchronously in the phase-1 function) already stops new
+        //    leases.
+        //
+        //    A drain timeout is *terminal* for this dispatch's outcome
+        //    metric: it records `TimedOut` and the subsequent hook
+        //    success/failure does NOT record a second outcome (one dispatch
+        //    = exactly one outcome). The hook still runs and its event /
+        //    returned outcome are unaffected — this is the contract the
+        //    removed outer `tokio::time::timeout` wrapper used to break.
+        let drain_result = managed.wait_for_in_flight_drain_erased(drain_timeout).await;
+        let drain_timed_out = drain_result.is_err();
+        if let Err(outstanding) = &drain_result {
+            if let Some(m) = &self.metrics {
+                m.record_slot_revoke_outcome(crate::metrics::SlotDispatchOutcome::TimedOut);
+            }
+            tracing::warn!(
+                outstanding = *outstanding,
+                "slot revoke: per-resource drain timed out; proceeding to \
+                 revoke hook (resource already tainted, no new leases)"
+            );
+        }
+
+        // 2. Dispatch the revoke hook against the live runtime, bounded by
+        //    the SAME per-resource budget. This is the only place the
+        //    budget can cut the tail short: a wedged `on_credential_revoke`
+        //    must not pin the fan-out row forever. A timed-out drain (above)
+        //    has *already* consumed the metric outcome, so a hook that then
+        //    also times out does not double-record.
+        let hook_outcome =
+            tokio::time::timeout(drain_timeout, managed.dispatch_on_revoke_erased(&slot)).await;
+        tracing::Span::current().record("duration_ms", tainted_at.elapsed().as_millis() as u64);
+
+        match hook_outcome {
+            Ok(Ok(())) => {
+                // Only record Success when the drain did not already record
+                // the terminal TimedOut outcome for this dispatch.
+                if !drain_timed_out && let Some(m) = &self.metrics {
+                    m.record_slot_revoke_outcome(crate::metrics::SlotDispatchOutcome::Success);
+                }
+                self.emit(ResourceEvent::SlotRevoked {
+                    key: key.clone(),
+                    slot: slot.clone(),
+                });
+                tracing::debug!("slot revoke hook completed");
+                RevokeTail::Done
+            },
+            Ok(Err(e)) => {
+                if !drain_timed_out && let Some(m) = &self.metrics {
+                    m.record_slot_revoke_outcome(crate::metrics::SlotDispatchOutcome::Failed);
+                }
+                self.emit(ResourceEvent::SlotRevokeFailed {
+                    key,
+                    slot,
+                    error: e.to_string(),
+                });
+                tracing::warn!(error = %e, "slot revoke hook failed");
+                RevokeTail::HookFailed(e)
+            },
+            Err(_elapsed) => {
+                // The hook itself wedged. The row stays tainted (phase 1).
+                // Record `TimedOut` unless the drain already did (one
+                // dispatch = exactly one outcome).
+                if !drain_timed_out && let Some(m) = &self.metrics {
+                    m.record_slot_revoke_outcome(crate::metrics::SlotDispatchOutcome::TimedOut);
+                }
+                self.emit(ResourceEvent::SlotRevokeFailed {
+                    key,
+                    slot,
+                    error: "revoke hook timed out".to_owned(),
+                });
+                tracing::warn!(
+                    timeout_ms = drain_timeout.as_millis() as u64,
+                    "slot revoke hook timed out (row stays tainted, no new leases)"
+                );
+                RevokeTail::HookTimedOut
+            },
+        }
+    }
+
+    /// Notifies a registered resource that one of its `#[credential]` slots
+    /// was revoked — **thin two-phase convenience** for non-fan-out callers
+    /// and tests.
+    ///
+    /// Equivalent to [`taint_slot`](Self::taint_slot) immediately followed by
+    /// [`drain_and_revoke`](Self::drain_and_revoke). The engine per-slot
+    /// rotation fan-out deliberately does **not** call this: it must run the
+    /// synchronous taint phase *outside* its `tokio::time::timeout` and wrap
+    /// only the awaited drain/hook tail, so a dropped timeout future can
+    /// never skip the taint (per-resource revoke deferral). This convenience is for the
+    /// no-timeout caller where the two phases run back-to-back on the same
+    /// task.
+    ///
+    /// # Errors
+    ///
+    /// - [`ErrorKind::NotFound`](crate::error::ErrorKind::NotFound) if no resource is registered for
+    ///   `key` at `scope`.
+    /// - [`ErrorKind::Ambiguous`](crate::error::ErrorKind::Ambiguous) if more than one
+    ///   resolved-credential row exists for `(key, scope)`.
+    /// - [`ErrorKind::Cancelled`](crate::error::ErrorKind::Cancelled) if the manager is shutting
+    ///   down.
+    /// - Whatever the resource's `on_credential_revoke` hook maps into [`Error`].
+    pub async fn revoke_slot(
+        &self,
+        key: &ResourceKey,
+        scope: ScopeLevel,
+        slot: &str,
+    ) -> Result<(), Error> {
+        let tainted = self.taint_slot(key, scope, slot)?;
+        self.drain_and_revoke(tainted, Self::DEFAULT_REVOKE_DRAIN_TIMEOUT)
+            .await
+            .into_result()
+    }
+
+    /// [`revoke_slot`](Self::revoke_slot) pinned to the **collision-free
+    /// structural** resolved per-slot credential identity — the
+    /// slot-identity-aware two-phase convenience.
+    ///
+    /// Equivalent to
+    /// [`taint_slot_for_identity`](Self::taint_slot_for_identity) immediately
+    /// followed by [`drain_and_revoke`](Self::drain_and_revoke); a
+    /// multi-tenant `(key, scope)` taints/drains/revokes the *specific*
+    /// resolved row instead of failing closed with
+    /// [`ErrorKind::Ambiguous`](crate::error::ErrorKind::Ambiguous). Like
+    /// [`revoke_slot`](Self::revoke_slot) this is the back-compat
+    /// back-to-back path; the engine fan-out drives the two phases separately
+    /// ([`taint_slot_for_identity`](Self::taint_slot_for_identity) outside
+    /// the timeout, then [`drain_and_revoke`](Self::drain_and_revoke)) per
+    /// per-resource revoke deferral.
+    ///
+    /// # Errors
+    ///
+    /// - [`ErrorKind::NotFound`](crate::error::ErrorKind::NotFound) if no row of `key` at `scope`
+    ///   matches `slot_identity`.
+    /// - [`ErrorKind::Cancelled`](crate::error::ErrorKind::Cancelled) if the manager is shutting
+    ///   down.
+    /// - Whatever the resource's `on_credential_revoke` hook maps into [`Error`].
+    pub async fn revoke_slot_for_identity(
+        &self,
+        key: &ResourceKey,
+        scope: ScopeLevel,
+        slot: &str,
+        slot_identity: &crate::dedup::SlotIdentity,
+    ) -> Result<(), Error> {
+        let tainted = self.taint_slot_for_identity(key, scope, slot, slot_identity)?;
+        self.drain_and_revoke(tainted, Self::DEFAULT_REVOKE_DRAIN_TIMEOUT)
+            .await
+            .into_result()
+    }
+
+    /// Type-erased `(key, scope)` → live `ManagedResource` resolution for
+    /// the slot-rotation entry points.
+    ///
+    /// `refresh_slot` / `revoke_slot` take a `ResourceKey` (not a generic
+    /// `R`), so they cannot use the typed `lookup::<R>`. This mirrors its
+    /// shutdown-race guard (reject once `shutting_down` is observed) and
+    /// resolves through the same registry the typed path uses, via the
+    /// type-erased `AnyManagedResource` view.
+    fn lookup_any_for_slot(
+        &self,
+        key: &ResourceKey,
+        scope: &ScopeLevel,
+    ) -> Result<Arc<dyn crate::registry::AnyManagedResource>, Error> {
+        use crate::registry::LookupOutcome;
+        self.shutdown_guard()?;
+        match self.registry.get(key, scope) {
+            LookupOutcome::Found(any) => Ok(any),
+            LookupOutcome::NotFound => Err(Error::not_found(key)),
+            // Fail closed: do not drive a rotation/revoke hook against an
+            // arbitrarily-chosen tenant's row when several resolved-
+            // credential rows share this `(key, scope)`. The engine's
+            // per-slot fan-out targets the specific resolved row.
+            LookupOutcome::Ambiguous { rows } => Err(Error::ambiguous(format!(
+                "{key}: {rows} resolved-credential registrations exist at this scope; \
+                 slot rotation/revoke must target a resolved row, not an ambiguous \
+                 (key, scope)"
+            ))
+            .with_resource_key(key.clone())),
+        }
+    }
+
+    /// Returns whether a registry row exists for
+    /// `(key, scope bag, slot_identity)`, keyed by the **collision-free
+    /// structural** resolved-credential identity.
+    ///
+    /// This is the engine-facing entry: the engine records a structural
+    /// [`SlotIdentity`](crate::dedup::SlotIdentity) at activation and asks
+    /// the same structural identity here, so a row is visible *only* under
+    /// its exact resolved binding set (no digest aliasing).
+    #[must_use]
+    pub fn has_registered_for_scope_identity(
+        &self,
+        key: &ResourceKey,
+        scope: &nebula_core::Scope,
+        slot_identity: &crate::dedup::SlotIdentity,
+    ) -> bool {
+        use crate::registry::AcquireLookupOutcome;
+        if self.shutdown_guard().is_err() {
+            return false;
+        }
+        matches!(
+            self.registry.get_acquire_for(key, scope, slot_identity),
+            AcquireLookupOutcome::Found { .. }
+        )
+    }
+
+    /// Returns whether a registry row exists for
+    /// `(key, scope level, slot_identity)`, keyed by the **collision-free
+    /// structural** resolved-credential identity.
+    ///
+    /// Prefer
+    /// [`has_registered_for_scope_identity`](Self::has_registered_for_scope_identity)
+    /// when the full scope bag is available (execution + org/workspace).
+    #[must_use]
+    pub fn has_registered_for_identity(
+        &self,
+        key: &ResourceKey,
+        scope: &ScopeLevel,
+        slot_identity: &crate::dedup::SlotIdentity,
+    ) -> bool {
+        let scope_bag = crate::context::minimal_scope_for_level(scope);
+        self.has_registered_for_scope_identity(key, &scope_bag, slot_identity)
+    }
+
+    /// [`lookup_any_for_slot`](Self::lookup_any_for_slot) pinned to a
+    /// resolved per-slot credential identity via
+    /// [`Registry::get_for`](crate::registry::Registry::get_for).
+    ///
+    /// [`get_for`](crate::registry::Registry::get_for) returns the
+    /// 2-variant [`PinnedLookup`](crate::registry::PinnedLookup): a
+    /// resolved slot identity pins exactly one `(scope, slot_identity)` row
+    /// by construction, so there is **no `Ambiguous` case to map** — the
+    /// "registry invariant breach" arm the old `u64` digest path had to
+    /// fabricate a fail-closed deny for is now type-unrepresentable.
+    fn lookup_any_for_slot_identity_structural(
+        &self,
+        key: &ResourceKey,
+        scope: &ScopeLevel,
+        slot_identity: &crate::dedup::SlotIdentity,
+    ) -> Result<Arc<dyn crate::registry::AnyManagedResource>, Error> {
+        use crate::registry::PinnedLookup;
+        self.shutdown_guard()?;
+        match self.registry.get_for(key, scope, slot_identity) {
+            PinnedLookup::Found(any) => Ok(any),
+            PinnedLookup::NotFound => Err(Error::not_found(key)),
+        }
+    }
+}
